fix(eos/frr)!: bad mapping between next-hop self/peer-address
Next-hop self is not the same than peer-address.

Next-hop self will be implemented properly for EOS/SONiC in a dedicated
PR as it will require a design doc. The implementation will be complicated because:
- In OpenConfig: next-hop self is set at the route-policy level
- In Juniper: like OpenConfig
- In EOS and SONiC(FRR): it is configurable only at the neighbor level

In the end, this commit does not change the current implementation in
AFK, because SELF is not authorized yet in the CMDB and Data Aggregation
API.