fix(junos): trying to clean route-map still being used
Before this commit, we were always deleting autogenerated route-map for
all SAFIs.

The purpose of the deletion is to start from a clean slate
when recreating a route-map (to avoid extra terms/options in the
route-map). The goal is not to remove extra policies, which must be done
in a dedicated step.

So this commit aims to only delete policy which will be recreated, to
avoid removing an autogenerated route-map which is still in use in the
actual configuration